refactor: rename schema files
renamed auth.schems to auth.schema
renamed feedback.schemas to feedback.schema
